requirements.ml.txt
jupyter
django-extensions
numpy>=1.11.2
pandas
scikit_surprise